requirements.txt
alembic==1.14.0
annotated-types==0.7.0
anyio==4.6.2.post1
asyncpg==0.30.0
certifi==2024.8.30
cfgv==3.4.0
click==8.1.7
coverage==7.6.7
distlib==0.3.9
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.5
fastapi-cli==0.0.5
filelock==3.16.1
flake8==7.1.1
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.0
identify==2.6.3
idna==3.10
iniconfig==2.0.0
isort==5.13.2
Jinja2==3.1.4
loguru==0.7.2
Mako==1.3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
mypy==1.13.0
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.2
platformdirs==4.3.6
pluggy==1.5.0
pre_commit==4.0.1
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycodestyle==2.12.1
pydantic==2.9.2
pydantic_core==2.23.4
pyflakes==3.2.0
Pygments==2.18.0
PyJWT==2.10.0
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==6.0.0
pytest-mock==3.14.0
pytest-order==1.3.0
python-dotenv==1.0.1
python-multipart==0.0.19
PyYAML==6.0.2
rich==13.9.4
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
typer==0.15.0
typing_extensions==4.12.2
uvicorn==0.32.1
uvloop==0.21.0
virtualenv==20.28.0
watchfiles==1.0.0
websockets==14.1
